git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/test@30635 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ casacore/casa/Containers/ValueHolderRep.cc
@@ -442,9 +442,9 @@ Int64 ValueHolderRep::asInt64() const
   case TpInt64:
     return itsInt64;
   case TpFloat:
-    return static_cast<Int64> (itsFloat);
+    return itsFloat;
   case TpDouble:
-    return static_cast<Int64> (itsDouble);
+    return itsDouble;
   default:
     ;
   }
